refactor(chat): move IM save action into chatActions
