+++ src/reactors/chat.js
@@ -1,8 +1,8 @@
 import { createSelector } from 'reselect'
 
-import { saveLocalChatMessages } from '../actions/chatMessageActions'
+import { saveLocalChatMessages, saveIMChatMessages } from '../actions/chatMessageActions'
 
-import { getLocalChat, getShouldSaveChat } from '../selectors/chat'
+import { getLocalChat, getIMChats, getShouldSaveChat } from '../selectors/chat'
 
 export const saveLocalChat = createSelector(
   [
@@ -32,71 +32,12 @@ export const saveLocalChat = createSelector(
 
 export const saveIMChat = createSelector(
   [
-    state => state.IMs,
+    getIMChats,
     getShouldSaveChat
   ],
-  (IMs, shouldSaveChat) => {
-    if (!shouldSaveChat || !IMs.some(chat => chat.get('hasUnsavedMSG'))) return null
+  (ims, shouldSaveChat) => {
+    if (!shouldSaveChat || !ims.some(chat => chat.get('hasUnsavedMSG'))) return null
 
-    return async (dispatch, getState, { hoodie }) => {
-      const unsavedChats = IMs.filter(chat => chat.get('hasUnsavedMSG'))
-
-      const chatsToSave = []
-
-      unsavedChats.forEach((chat, key) => {
-        const messages = chat.get('messages')
-
-        const toSaveMsg = messages.filter(msg => !msg.get('didSave')).toJSON().map(msg => {
-          const toSave = Object.assign({}, msg)
-          delete toSave.didSave
-          return toSave
-        })
-
-        chatsToSave.push(...toSaveMsg)
-      })
-
-      dispatch({
-        type: 'StartSavingIMMessages',
-        chats: chatsToSave.reduce((all, msg) => {
-          let messages = all[msg.chatUUID]
-
-          if (messages == null) {
-            messages = []
-            all[msg.chatUUID] = messages
-          }
-
-          messages.push(msg._id)
-          return all
-        }, {})
-      })
-
-      const saved = await hoodie.cryptoStore.updateOrAdd(chatsToSave)
-
-      const results = saved.reduce((all, msg, index) => {
-        const chatUUID = chatsToSave[index].chatUUID
-        let chat = all[chatUUID]
-
-        if (chat == null) {
-          chat = {
-            saved: [],
-            didError: []
-          }
-          all[chatUUID] = chat
-        }
-
-        if (msg instanceof Error) {
-          chat.didError.push(chatsToSave[index]._id)
-        } else {
-          chat.saved.push(msg)
-        }
-
-        return all
-      }, {})
-
-      dispatch({
-        type: 'didSaveIMMessages',
-        chats: results
-      })
-    }
+    return saveIMChatMessages(ims)
   }
 )
