+++ src/selectors/chat.js
@@ -2,6 +2,8 @@ import { createSelector } from 'reselect'
 
 export const getLocalChat = state => state.localChat
 
+export const getIMChats = state => state.IMs
+
 export const getShouldSaveChat = createSelector(
   [
     state => state.account
